Add article edit function

+++ singo/service/article/delete_article_service.go
@@ -3,6 +3,7 @@ package service
 import (
 	"singo/model"
 	"singo/serializer"
+	"fmt"
 )
 
 // DeleteArticleService 删除投稿的服务
@@ -10,7 +11,7 @@ type DeleteArticleService struct {
 }
 
 // Delete 删除文章
-func (service *DeleteArticleService) Delete(id string) serializer.Response {
+func (service *DeleteArticleService) Delete(user string, id string) serializer.Response {
 	var Article model.Article
 	err := model.DB.First(&Article, id).Error
 	if err != nil {
@@ -20,7 +21,16 @@ func (service *DeleteArticleService) Delete(id string) serializer.Response {
 			Error:  err.Error(),
 		}
 	}
-
+	fmt.Println(user)
+	fmt.Println(Article.UserName)
+	fmt.Println(user != Article.UserName)
+	if user != Article.UserName {
+		return serializer.Response{
+			Status: 404,
+			Msg:    "没有权限删除",
+			Error:  "no permission",
+		}
+	}
 	err = model.DB.Delete(&Article).Error
 	if err != nil {
 		return serializer.Response{
